+++ server/index.js
@@ -3,22 +3,14 @@ const axios = require('axios');
 const path = require('path');
 const config = require('../config.js');
 const { Pool } = require('pg');
+const db = require('../db');
+const reviews = require('./reviews');
 
 const PORT = 3002;
 const app = express();
-const router = express.Router();
-app.use(express.json());
 
-const db = new Pool({
-  host: config.host,
-  user: config.user,
-  password: config.password,
-  database: config.database
-});
+app.use('/reviews', reviews);
 
-db.connect()
-  .then( () =>  console.log(`Connected to database ${config.database}`))
-  .catch( (err) => console.log('Error connecting to db: ', err));
 
 app.listen(PORT, () => {
   console.log(`Server listening at localhost:${PORT}!`);
